// web-app/src/interfaces/hr/recruiting/planning.ts
export interface IRecruitingPlanning {
    id: number;
    departmentId: number;
    departmentTitle: string;
    postTitle: string;
    postInfo: string;
    education: string;
    experience: string;
    salariesRange: string;
    technology: string;
    status: string;
    reason: string;
    hrattacheId: number;
    hrAttacheName: string;
    createUserName: string;
    createBy: number;
    createDate: string;
    updateUserName: string;
    updateBy: number;
    updateDate: string;
}